+++ HISTORY.md
@@ -2,7 +2,8 @@
 
 ## Version 1.0.25 - 7 September 2026
 1. Update buttons
-2.
+2. Bind mouewheel to frm_compare
+3.
 
 ## Version 1.0.24 - 31 August 2026
 1. Sort projects in tree by name.lower()